+++ tests/test_utils.py
@@ -1,4 +1,6 @@
-from pomp.core.utils import iterator
+import pytest
+
+from pomp.core.utils import iterator, Planned, CancelledError, NotDoneYetError
 
 
 def test_iterator():
@@ -9,3 +11,35 @@ def test_iterator():
     assert hasattr(iterator(1), '__iter__')
 
     assert hasattr(iterator(iterator('b')), '__iter__')
+
+
+def test_planned():
+
+    planned = Planned()
+
+    assert not planned.cancelled()
+    assert not planned.done()
+
+    with pytest.raises(NotDoneYetError):
+        planned.result()
+
+    def cb1(result):
+        pass
+
+    def cb2(result):
+        raise RuntimeError("some test exception")
+
+    planned.add_done_callback(cb1)
+    planned.add_done_callback(cb2)
+
+    planned.set_result('ok')
+
+    assert not planned.cancelled()
+    assert planned.done()
+    assert planned.result() == 'ok'
+
+    planned = Planned()
+    planned.cancel()
+
+    with pytest.raises(CancelledError):
+        planned.result()